Mac AVFoundation Video Capture: float frame rate
Floating point frame rate is supported in Chromium code now and is
also provided by AVFoundation, eliminate the unnecessary cast to int.

BUG=380463

Review URL: https://codereview.chromium.org/365953003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281260 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e1f462095395e3790c20980c4ce2bad15d9c0b5d

+++ media/video/capture/mac/video_capture_device_avfoundation_mac.mm
@@ -74,7 +74,7 @@
            [format videoSupportedFrameRateRanges]) {
       media::VideoCaptureFormat format(
           gfx::Size(dimensions.width, dimensions.height),
-          static_cast<int>(frameRate.maxFrameRate),
+          frameRate.maxFrameRate,
           pixelFormat);
       formats->push_back(format);
       DVLOG(2) << name.name() << " resolution: "
